Move ImageSession to a separate module
