+++ naturtag/storage/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa: F401
 from naturtag.storage.app_state import AppState
-from naturtag.storage.client import ImageSession, iNatDbClient
+from naturtag.storage.client import iNatDbClient
+from naturtag.storage.session import ImageSession
 from naturtag.storage.settings import Settings
 from naturtag.storage.setup import setup
